refactor(win32): move compat.rs
