+++ winio-ui-app-kit/Cargo.toml
@@ -43,7 +43,7 @@ objc2-web-kit = { version = "0.3", optional = true }
 media = ["dep:objc2-av-kit", "dep:objc2-av-foundation", "dep:objc2-core-media"]
 webview = ["dep:objc2-web-kit"]
 
-compio-compat = ["compio/compat"]
+compio-compat = ["compio/runtime", "compio/compat"]
 
 objc-static = [
     "objc2/unstable-static-class-inlined",
