+++ winio-ui-win32/Cargo.toml
@@ -14,6 +14,7 @@ winio-callback = { workspace = true, optional = true }
 winio-pollable = { workspace = true }
 winio-ui-windows-common = { workspace = true, features = ["win32"] }
 
+compio = { workspace = true, default-features = false, optional = true }
 compio-log = { workspace = true }
 
 arrayvec = { workspace = true }
@@ -61,7 +62,7 @@ webview = [
     "dep:windows-core",
     "dep:local-sync",
 ]
-compio-compat = ["winio-ui-windows-common/compio-compat"]
+compio-compat = ["compio/runtime", "compio/compat"]
 
 dark-mode = ["winio-ui-windows-common/dark-mode"]
