+++ winio-ui-win32/src/lib.rs
@@ -5,8 +5,6 @@
 #![cfg_attr(feature = "once_cell_try", feature(once_cell_try))]
 #![cfg(windows)]
 
-#[cfg(feature = "compio-compat")]
-pub use winio_ui_windows_common::CompioAdapter;
 pub use winio_ui_windows_common::{Error, Result};
 
 mod runtime;
@@ -14,3 +12,8 @@ pub use runtime::*;
 
 mod ui;
 pub use ui::*;
+
+#[cfg(feature = "compio-compat")]
+mod compat;
+#[cfg(feature = "compio-compat")]
+pub use compat::*;
